Following better organization principles
